feature/ added state machine for shooter
Added a state machine for the shooter where if you can select the shooting mode and it will start preparing the wheels to shoot. Then the driver can command it to shoot it. The indexer also stops intaking when it detects a beam brake. The intaking and outtaking runs both the indexer and the intake rollers.